% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix_from_list.R
\name{adjacency_matrix_from_list}
\alias{adjacency_matrix_from_list}
\title{Create an adjacency matrix from a list of complexes}
\usage{
adjacency_matrix_from_list(complexes)
}
\arguments{
\item{complexes}{a list of complexes, with each entry containing complex
subunits as a character vector}
}
\value{
an adjacency matrix between all complex subunits
}
\description{
Convert a list of complexes into a pairwise adjacency matrix. The resulting
square adjacency matrix contains ones for proteins that are found in the
same complex and zeroes otherwise.
}
\examples{
data(gold_standard)
adj = adjacency_matrix_from_list(data)

}
