% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_chromatogram_replicates.R
\name{average_chromatogram_replicates}
\alias{average_chromatogram_replicates}
\title{Average values of chromatogram matrix over several replicates}
\usage{
average_chromatogram_replicates(mats, min_replicates = 1)
}
\arguments{
\item{mats}{a list of chromatogram matricies}

\item{min_replicates}{the minimum number of replicates the protein fraction
must appear in.}
}
\value{
a matrix of averaged chromatogram values between replicates
}
\description{
Find all proteins that are detected between all replicates and generates
the mean value of each fraction of each protein that occur in at least
a certain number of replicates, adding \code{NA}s otherwise.
}
