% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_precision.R
\name{calculate_precision}
\alias{calculate_precision}
\title{Calculate precision at each point in a sequence}
\usage{
calculate_precision(labels)
}
\arguments{
\item{labels}{a vector of zeroes (FPs) and ones (TPs)}
}
\value{
a vector of the same length giving the precision at each point in the
input vector
}
\description{
Calculate the precision of a list of interactions at each point in the list,
given a set of labels.
}
