% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_gaussians.R
\name{choose_gaussians}
\alias{choose_gaussians}
\title{Fit a mixture of Gaussians to a chromatogram curve and choose the 
best-fitting model}
\usage{
choose_gaussians(chromatogram, max_gaussians = 5, criterion = c("AICc",
  "AIC", "BIC"))
}
\arguments{
\item{chromatogram}{a numeric vector corresponding to the chromatogram trace}

\item{max_gaussians}{the maximum number of Gaussians to fit; defaults to 5.
Note that Gaussian mixtures with more parameters than observed (i.e., 
non-zero or NA) points will not be fit.}

\item{criterion}{the criterion to use for model selection;
one of "AICc" (corrected AIC, and default), "AIC", or "BIC"}
}
\value{
a list with three entries: the number of Gaussians used to fit
the curve in the best model, the R^2 of that fit, and the fit object 
generated by \code{nls}
}
\description{
Fit mixtures of one or more Gaussians to the curve formed by a chromatogram
profile, and choose the best fitting model using an information criterion
of choice.
}
