% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_features.R
\name{concatenate_features}
\alias{concatenate_features}
\title{Combine features across multiple replicates}
\usage{
concatenate_features(feature_list)
}
\arguments{
\item{feature_list}{a list of feature data frames, as produced by 
\code{\link[PrInCE]{calculate_features}}, with proteins in the first two 
columns}
}
\value{
a data frame containing features for all protein pairs across all 
  replicates
}
\description{
Concatenate features extracted from multiple replicates to a single data 
frame that will be used as input to a classifier. Doing so allows the 
classifier to naturally weight evidence for an interaction between each
protein pair from each feature in each replicate in proportion to its
discriminatory power on known examples.
}
