% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix_from_data_frame.R
\name{feature_matrix_from_data_frame}
\alias{feature_matrix_from_data_frame}
\title{Create a feature matrix for a naive Bayes classifier from a data frame}
\usage{
feature_matrix_from_data_frame(dat, profile_matrix, col_name)
}
\arguments{
\item{dat}{a data frame with pairwise interactions in the first two columns,
and optionally some information about each interaction (such as a score or
category) in the third column}

\item{profile_matrix}{the profile matrix for which interactions are being 
predicted}

\item{col_name}{the name of the column in the input data frame that contains
interaction features}
}
\value{
a square matrix with the same row and column names as the input 
profile matrix, for use in interaction prediction
}
\description{
Convert a data frame containing pairwise interactions, and  
a score or other data associated with each interaction, into a feature matrix
with the same dimensions and row/column names as those used to train 
naive Bayes classifiers for default PrInCE features.
}
