% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix_from_expression.R
\name{feature_matrix_from_expression}
\alias{feature_matrix_from_expression}
\title{Create a feature matrix for a naive Bayes classifier from expression data}
\usage{
feature_matrix_from_expression(expr, profile_matrix, ...)
}
\arguments{
\item{expr}{a matrix containing gene or protein expression data, with 
genes/proteins in columns and samples in rows}

\item{profile_matrix}{the profile matrix for which interactions are being 
predicted}

\item{...}{arguments passed to \code{cor}}
}
\value{
a square matrix with the same row and column names as the input 
profile matrix, for use in interaction prediction
}
\description{
Convert a gene or protein expression matrix into a feature matrix
with the same dimensions and row/column names as those used to train 
naive Bayes classifiers for default PrInCE features by calculating
the correlation between each pair of genes or proteins.
}
