% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fit_model}
\alias{fit_model}
\title{Fit the mixture of Gaussians used by the \code{fold_change} function}
\usage{
fit_model(df, fit_condition = F, min_gaussians = 1,
  max_gaussians = 5, max_iterations = 10, criterion = c("AICc",
  "AIC", "BIC"), min_R_squared = 0.5, filter_gaussians_center = T,
  filter_gaussians_height = 0.15, filter_gaussians_variance_min = 0.1,
  filter_gaussians_variance_max = 50, sigma_default = 2)
}
\arguments{
\item{df}{a data frame containing the variables \code{fraction}, 
\code{ratio}, and \code{condition}}

\item{fit_condition}{whether to include condition as a covariate in the model}

\item{min_gaussians}{the inimum number of Gaussians to fit; defaults to 1.}

\item{max_gaussians}{the maximum number of Gaussians to fit; defaults to 5.
Gaussian mixtures with more parameters than observed (i.e., 
non-zero or NA) points will not be fit.}

\item{max_iterations}{the number of times to try fitting the curve with
different initial conditions; defaults to 10}

\item{criterion}{the criterion to use for model selection;
one of \code{AICc} (corrected AIC, and default), \code{AIC}, or \code{BIC}}

\item{min_R_squared}{the minimum R-squared value to accept when fitting the
curve with different initial conditions; defaults to 0.5}

\item{filter_gaussians_center}{true or false: filter Gaussians whose centres
fall outside the bounds of the chromatogram}

\item{filter_gaussians_height}{Gaussians whose heights are below this 
fraction of the chromatogram height will be filtered. Setting this value to
zero disables height-based filtering of fit Gaussians}

\item{filter_gaussians_variance_min}{Gaussians whose variance falls below 
this number of fractions will be filtered. Setting this value to
zero disables filtering.}

\item{filter_gaussians_variance_max}{Gaussians whose variance is above
this number of fractions will be filtered. Setting this value to
zero disables filtering.}

\item{sigma_default}{the default mean initial value of sigma}
}
\description{
Fit the mixture of Gaussians used by the \code{fold_change} function
}
