% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change}
\alias{fold_change}
\title{Identify significantly changing proteins using a likelihood ratio test}
\usage{
fold_change(chrom1, chrom2)
}
\arguments{
\item{chrom1}{the chromatogram of the protein of interest in the first 
condition (e.g., the M/L channel)}

\item{chrom2}{the chromatogram of the protein of interest in the first 
condition (e.g., the H/L channel)}
}
\description{
This function identifies proteins with a significant change between two
conditions in a protein correlation profiling experiment, by fitting two 
different mixtures of Gaussians: one with the condition as a covariate, and
one without. The two models are then compared with a likelihood ratio test
to evaluate whether taking condition into account provides a significantly 
better fit.
}
