% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_aic.R
\name{gaussian_aic}
\alias{gaussian_aic}
\title{Calculate AIC for a Gaussian mixture model}
\usage{
gaussian_aic(coefs, chromatogram)
}
\arguments{
\item{coefs}{the coefficients of the Gaussian mixture model, output by 
\code{\link{fit_gaussians}}}

\item{chromatogram}{the raw elution profile}
}
\value{
the AIC of the fit model
}
\description{
Calculates the AIC for a curve fit with a Gaussian mixture model
by nonlinear least squares optimization. This function permits the 
calculation of the AIC after rejecting some Gaussians in the model, 
for example because their centres are outside the bounds of the profile.
}
