% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_aicc.R
\name{gaussian_aicc}
\alias{gaussian_aicc}
\title{Calculate the corrected AIC for a Gaussian mixture model}
\usage{
gaussian_aicc(fit, indices)
}
\arguments{
\item{fit}{the curve fit by the \code{stats::nls} function}
}
\value{
the corrected AIC of the fit model
}
\description{
Calculates the corrected AIC for a curve fit with a Gaussian mixture model
by nonlinear least squares optimization.
}
