% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_bic.R
\name{gaussian_bic}
\alias{gaussian_bic}
\title{Calculate BIC for a Gaussian mixture model}
\usage{
gaussian_bic(coefs, chromatogram)
}
\arguments{
\item{coefs}{the coefficients of the Gaussian mixture model, output by 
\code{\link{fit_gaussians}}}

\item{chromatogram}{the raw elution profile}
}
\value{
the BIC of the fit model
}
\description{
Calculates the BIC for a curve fit with a Gaussian mixture model
by nonlinear least squares optimization. This function permits the 
calculation of the BIC after rejecting some Gaussians in the model, 
for example because their centres are outside the bounds of the profile.
}
