% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_standard.R
\docType{data}
\name{gold_standard}
\alias{gold_standard}
\title{Reference set of human protein complexes}
\format{
a list containing 467 entries (character vectors)
}
\source{
\url{https://www.ebi.ac.uk/complexportal/complex/organisms}
}
\usage{
data(gold_standard)
}
\description{
A reference set of 467 experimentally confirmed human protein complexes,
derived from the EBI Complex Portal database.
}
\details{
467 protein complexes, ranging in size from 2 to 44 proteins and involving
877 proteins in total, to provide a reference set of true positive and true 
negative interactions (intra- and inter-complex interactions, respectively)
for demonstration in PrInCE analysis of a co-elution dataset.
Other "gold standards" are possible in practice, most notably the CORUM
database; however, the Complex Portal reference set is included in this
package due to its CC-BY license.
}
\keyword{datasets}
