% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_neighbors.R
\name{impute_neighbors}
\alias{impute_neighbors}
\title{Impute single missing values}
\usage{
impute_neighbors(chromatogram)
}
\arguments{
\item{chromatogram}{a numeric vector corresponding to the chromatogram trace}
}
\value{
the imputed chromatogram
}
\description{
Impute single missing values within a chromatogram profile as the average
of their neighbors.
}
\examples{
data(scott)
chrom <- scott[16, ]
imputed <- impute_neighbors(chrom)

}
