% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unweighted.R
\name{is_unweighted}
\alias{is_unweighted}
\title{Test whether a network is unweighted}
\usage{
is_unweighted(network)
}
\arguments{
\item{network}{the network to analyze}
}
\value{
true if the input network is a square logical or numeric matrix
}
\description{
Test whether a network is unweighted
}
