% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_feature_from_data_frame.R
\name{make_feature_from_data_frame}
\alias{make_feature_from_data_frame}
\title{Create a feature vector for a classifier from a data frame}
\usage{
make_feature_from_data_frame(dat, target, col_name, default_value = NA)
}
\arguments{
\item{dat}{a data frame with pairwise interactions in the first two columns,
and the features to convert to a vector in some other column}

\item{target}{the data frame of features to be used by the classifier, 
with protein pairs in the first two columns}

\item{col_name}{the name of the column in the first data frame that 
contains interaction features}

\item{default_value}{the default value for protein pairs that are not in the 
first data frame (set, by default, to \code{NA})}
}
\value{
a vector matching the dimensions and order of the feature data frame,
to use as input for a classifier in interaction prediction
}
\description{
Convert a data frame containing pairwise interactions, and  
a score or other data associated with each interaction, into a feature vector
that matches the dimensions of a data frame used as input to a classifier, 
such as a naive Bayes, random forests, or support vector machine classifier.
}
