% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels}
\alias{make_labels}
\title{Make labels for a classifier based on a gold standard}
\usage{
make_labels(gold_standard, input)
}
\arguments{
\item{gold_standard}{an adjacency matrix of gold-standard interactions}

\item{input}{a data frame with interacting proteins in the first two
columns}
}
\value{
a vector of the same length as the input dataset, containing 
\code{NA}s for protein pairs not in the gold standard and ones or zeroes
based on the content of the adjacency matrix
}
\description{
Create labels for a classifier for protein pairs in the same order as 
in a dataset that will be used as input to a classifier, in a 
memory-friendly way.
}
