% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_aicc.R
\name{nls_aicc}
\alias{nls_aicc}
\title{Calculate the corrected AIC for a NLS fit}
\usage{
nls_aicc(fit)
}
\arguments{
\item{fit}{the curve fit by the \code{stats::nls} function}
}
\value{
the corrected AIC of the fit model
}
\description{
Calculates the corrected AIC for a curve fit with a Gaussian mixture model
by nonlinear least squares optimization, using the base R \code{nls}
function.
}
