% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_SVM_ensemble.R
\name{predict_SVM_ensemble}
\alias{predict_SVM_ensemble}
\title{Predict interactions using an ensemble of support vector machines}
\usage{
predict_SVM_ensemble(input, labels, models = 1, cv_folds = 10,
  seed = 0)
}
\arguments{
\item{input}{a data frame containing interacting gene/protein pairs in the
first two columns, and the features to use for classification in the 
remaining columns}

\item{labels}{labels for each interaction in \code{input}: 0 for negatives,
1 for positives, and NA for interactions outside the reference set}

\item{models}{the number of support vector machines to train}

\item{cv_folds}{the number of folds to split the reference dataset into 
when training each support vector machine. By default, each 
classifier uses ten-fold cross-validation, i.e., the classifier is trained
on 90\% of the dataset and used to classify the remaining 10\%}

\item{seed}{the seed for the random number generator, used to ensure
reproducibility}
}
\value{
the input data frame of pairwise interactions, ranked by the 
median of classifier scores across all ensembled models
}
\description{
Use an ensemble of support vector machines to predict interactions from
co-elution dataset features. The ensemble approach ensures that 
results are robust to the partitioning of the dataset into folds. For each
model, the median of classifier scores across all folds is calculated.
Then, the median of all such medians across all models is calculated.
}
