% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix_from_data_frame.R
\name{adjacency_matrix_from_data_frame}
\alias{adjacency_matrix_from_data_frame}
\title{Create an adjacency matrix from a data frame}
\usage{
adjacency_matrix_from_data_frame(dat)
}
\arguments{
\item{dat}{a data frame with pairwise interactions in the first two columns}
}
\value{
an adjacency matrix between all interacting proteins
}
\description{
Convert a data frame containing pairwise interactions into an
adjacency matrix. The resulting square adjacency matrix contains ones for 
proteins that are found in interactions and zeroes otherwise.
}
