% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_label_matrix.R
\name{make_label_matrix}
\alias{make_label_matrix}
\title{Create a gold-standard label matrix}
\usage{
make_label_matrix(gold_standard, profile_matrix)
}
\arguments{
\item{gold_standard}{an adjacency matrix of gold-standard interactions}

\item{profile_matrix}{the profile matrix for which interactions are being
predicted}
}
\value{
a matrix with the same dimensions as the input feature matrix
being used as input to the classifier
}
\description{
Create a gold-standard label matrix with the same dimensions as the 
feature matrices being classified.
}
