% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_interactions.R
\name{predict_interactions}
\alias{predict_interactions}
\title{Predict interactions for a profile matrix}
\usage{
predict_interactions(profile_matrix, gaussians, gold_standard)
}
\arguments{
\item{profile_matrix}{a numeric matrix of co-elution profiles, with proteins
in rows}

\item{gaussians}{a list of Gaussian mixture models fit to the profile matrix
by \code{link{build_gaussians}}}

\item{gold_standard}{a matrix or data frame of "gold standard" interactions
used to train a naive Bayes classifier}
}
\value{
a data frame containing the values of these five features for each
protein pair and the score output by the naive Bayes classifier
}
\description{
Predict pairwise interactions between proteins in a profile matrix, using a
naive Bayes classifier trained on dataset-derived features including:
* one minus the Pearson correlation coefficient and 
* the corresponding P value, 
* the Euclidean distance between profiles, 
* the difference (in fractions) between the maximum values of each profile, 
(the peak score) and
* the co-apex score, defined as the minimum Euclidean distance between any
two Gaussians
}
