% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_precision.R
\name{threshold_precision}
\alias{threshold_precision}
\title{Threshold interactions at a given precision cutoff}
\usage{
threshold_precision(interactions, threshold)
}
\arguments{
\item{interactions}{the ranked list of interactions output by 
\code{\link[PrInCE]{predict_interactions}}, including a \code{precision} 
column}

\item{threshold}{the minimum precision of the unweighted interaction 
network to return}
}
\value{
the subset of the original ranked list at the given precision
}
\description{
Threshold interactions at a given precision cutoff
}
